#11: Addendum 816652d

+++ src/main/java/com/tmarsteel/jcli/util/formatting/Indentation.java
@@ -8,7 +8,7 @@ public class Indentation
     private final char indentationChar;
     private final int indentationAmount;
 
-    public Indentation(short indentationAmount) {
+    public Indentation(int indentationAmount) {
         this(' ', indentationAmount);
     }
